Cleaning and re-organizing project which closes issue #65
Removed unnecessary files (such as setAuthorizationToken.js), commented code on most pages, reduced some css clutter on login/signup forms, and properly organized react components/pages to better follow coding convetion (pages with compoonents injected into them)
