+++ client/src/App.css
@@ -28,8 +28,7 @@
 }
 
 .App-header {
-  /*background-color: #282c34;*/
-  background-image: url(./windowsbackground.jpg);
+  background-color: #282c34;
   min-height: 100vh;
   display: flex;
   flex-direction: column;
@@ -39,8 +38,12 @@
   color: black;
 }
 
-.Windows-background {
-  background-image: url(./windowsbackground.jpg);
+form {
+  color: white;
+}
+
+p {
+  color: black;
 }
 
 .App-link {
@@ -59,7 +62,7 @@
 h1 {
   color: #111;
   font-family: "Helvetica Neue", sans-serif;
-  font-size: 100px;
+  font-size: calc(10px + 2vmin);
   font-weight: bold;
   text-decoration: underline;
   letter-spacing: -1px;
