+++ client/src/App.js
@@ -3,7 +3,7 @@ import { Route, Switch, BrowserRouter as Router } from "react-router-dom";
 import LoginSignupPage from "./pages/loginsignup";
 import HomePage from "./pages/homepage";
 import FollowingPage from "./pages/managefollowing";
-import ProfilePage from "./components/profile";
+import ProfilePage from "./pages/profilepage";
 import "./App.css";
 
 function App() {
