+++ client/src/components/signup.js
@@ -1,9 +1,9 @@
 import React, { Component } from "react";
-import { Form, Button } from "react-bootstrap";
+import { Form, Button, Card } from "react-bootstrap";
 import { Redirect } from "react-router-dom";
 import axios from "axios";
 import auth from "../authentication/auth-login";
-import "../App.css";
+import "../styling/signupstyle.css";
 
 class RenderSignUpForm extends Component {
   constructor(props) {
@@ -61,46 +61,61 @@ class RenderSignUpForm extends Component {
       return <Redirect to="/" />;
     }
     return (
-      <div className="App">
-        <header className="App-header">
-          <h1>Sign Up</h1>
-          <br />
-          <Form onSubmit={this.onSubmit}>
-            <Form.Group controlId="email">
-              <Form.Label>Email</Form.Label>
-              <Form.Control
-                type="email"
-                placeholder="Enter email"
-                name="email"
-                value={this.state.email}
-                onChange={this.onChangeEmail}
-              />
-            </Form.Group>
-            <Form.Group controlId="username">
-              <Form.Label>Username</Form.Label>
-              <Form.Control
-                type="username"
-                placeholder="Enter username"
-                name="username"
-                value={this.state.username}
-                onChange={this.onChangeUsername}
-              />
-            </Form.Group>
-            <Form.Group controlId="password">
-              <Form.Label>Password</Form.Label>
-              <Form.Control
-                type="password"
-                placeholder="Password"
-                name="password"
-                value={this.state.password}
-                onChange={this.onChangePassword}
-              />
-            </Form.Group>
-            <Button variant="primary" type="submit">
-              Submit
-            </Button>
-          </Form>
-        </header>
+      <div className="Signup">
+        <div className="Signup-header">
+          <Card
+            border="dark"
+            className="rounded mb-20"
+            bg="light"
+            style={{ width: "25rem", padding: "40px" }}
+            text="dark"
+          >
+            <h1>Create an account</h1>
+            <Card.Body>
+              <Form onSubmit={this.onSubmit}>
+                <Form.Group controlId="email">
+                  <Card.Text>
+                    <Form.Label>Enter your Email</Form.Label>
+                  </Card.Text>
+                  <Form.Control
+                    type="email"
+                    placeholder="Enter email"
+                    name="email"
+                    value={this.state.email}
+                    onChange={this.onChangeEmail}
+                  />
+                </Form.Group>
+                <Form.Group controlId="username">
+                  <Card.Text>
+                    <Form.Label>Create a Username</Form.Label>
+                  </Card.Text>
+                  <Form.Control
+                    type="username"
+                    placeholder="Enter username"
+                    name="username"
+                    value={this.state.username}
+                    onChange={this.onChangeUsername}
+                  />
+                </Form.Group>
+                <Form.Group controlId="password">
+                  <Card.Text>
+                    <Form.Label>Enter a Password</Form.Label>
+                  </Card.Text>
+                  <Form.Control
+                    type="password"
+                    placeholder="Password"
+                    name="password"
+                    value={this.state.password}
+                    onChange={this.onChangePassword}
+                  />
+                </Form.Group>
+                <Button variant="primary" type="submit">
+                  Create
+                </Button>
+              </Form>
+            </Card.Body>
+          </Card>
+        </div>
       </div>
     );
   }
