+++ client/src/styling/loginstyle.css
@@ -1,0 +1,25 @@
+.Login {
+  text-align: center;
+}
+
+.Login-header {
+  background-color: #282c34;
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  font-size: calc(10px + 2vmin);
+  color: #111;
+}
+
+h1 {
+  color: #111;
+  font-family: "Helvetica Neue", sans-serif;
+  font-size: calc(10px + 2vmin);
+  font-weight: bold;
+  text-decoration: none;
+  letter-spacing: -1px;
+  line-height: 1;
+  text-align: center;
+}
