+++ server/routes/following.js
@@ -3,38 +3,48 @@ const express = require("express");
 const router = new express.Router();
 const auth = require("../auth");
 
-//Send all users that are followed for a requesting user
+//Path to return all of a specific users followers.
 router.get("/", auth, async (req, res) => {
   try {
+    //Search the database for the user who is REQUESTING to view all his followers.
+    //If this user does not exist, return status 400.
     const user = await User.findById(req.user.id);
     if (!user) {
       res.status(400).json({ message: "User with matching ID not found" });
     }
 
+    //If the REQUESTING user exists, but the user does not have any users who they
+    //are following, return status 400.
     const userFollowing = user.following;
     if (!userFollowing) {
       res.status(400).json({ message: "User Followers not found" });
     }
-    console.log({ following: userFollowing });
+
+    //If the REQUESTING user exists, and they are following other users, return the list
+    //of all their followers.
     res.status(200).json({ following: userFollowing });
   } catch (error) {
     res.status(400).json({ message: error });
   }
 });
 
-//Follow a new User
+//Path to follow a new user.
 router.post("/new", auth, async (req, res) => {
   try {
-    //Assigning id of user where the new follow will be stored.
+    //Defining the ID of the user who is REQUESTING to follow a new user.
     const userId = req.user.id;
 
-    //Check if a username was sent with the request.
+    //If the username of the REQUESTED follow does not exist in the HTTP request body,
+    //return status 404 not found.
     if (!req.body.username) {
       return res.status(404).json({ message: "No username received." });
     }
-    //Find id associated with that username
+
+    //Initialize followUsername with the username of the REQUESTED follow.
     const followUsername = req.body.username;
-    //Check if user exists inside database. If user does not exist, return status 404.
+
+    //Check if the REQUESTED follow actually exists inside of the database.
+    //If the user does not exist, return status 400.
     let userToFollow = await User.findOne({ username: followUsername });
     if (!userToFollow) {
       return res.status(400).json({
@@ -42,7 +52,8 @@ router.post("/new", auth, async (req, res) => {
       });
     }
 
-    //Update the user's document with the new follow
+    //Update the REQUESTING user's account with the new user who they are following using
+    //$addToSet (mongoose syntax). If successful, return status 200.
     await User.findByIdAndUpdate(
       userId,
       {
